# contracts/openapi/api-v1-v0.6.0.json
{
  "components": {
    "schemas": {
...
      "CustomerCreate": {
        "additionalProperties": false,
        "properties": {
          "display_name": {
            "anyOf": [
              {
                "maxLength": 128,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "merchant_customer_reference": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Merchant Customer Reference",
            "type": "string"
          }
        },
        "required": [
          "merchant_customer_reference"
        ],
        "title": "CustomerCreate",
        "type": "object"
      },
      "EmptyCommand": {
        "additionalProperties": false,
        "properties": {},
        "title": "EmptyCommand",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PaymentIntentCreate": {
        "additionalProperties": false,
        "properties": {
          "amount": {
            "exclusiveMinimum": 0.0,
            "maximum": 9.223372036854776e+18,
            "title": "Amount",
            "type": "integer"
          },
          "currency": {
            "const": "INR",
            "default": "INR",
            "title": "Currency",
            "type": "string"
          },
          "customer_id": {
            "pattern": "^cus_[0-9a-f]{32}$",
            "title": "Customer Id",
            "type": "string"
          },
          "merchant_reference": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Merchant Reference",
            "type": "string"
          }
        },
        "required": [
          "customer_id",
          "merchant_reference",
          "amount"
        ],
        "title": "PaymentIntentCreate",
        "type": "object"
      },
...
      "RefundCreate": {
        "additionalProperties": false,
        "properties": {
          "amount": {
            "exclusiveMinimum": 0.0,
            "maximum": 9.223372036854776e+18,
            "title": "Amount",
            "type": "integer"
          },
          "currency": {
            "const": "INR",
            "default": "INR",
            "title": "Currency",
            "type": "string"
          },
          "merchant_refund_reference": {
            "anyOf": [
              {
                "maxLength": 128,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Merchant Refund Reference"
          }
        },
        "required": [
          "amount"
        ],
        "title": "RefundCreate",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Synthetic-data-only RelayPay merchant and operator API. Never submit real payment, bank-account, identity, or customer data.",
    "title": "RelayPay API",
    "version": "0.6.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/customers": {
      "post": {
        "operationId": "post_customer_api_v1_customers_post",
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CustomerCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Customer",
        "tags": [
          "payments"
        ]
      }
    },
    "/api/v1/operations/{operation_id}": {
      "get": {
        "operationId": "get_operation_api_v1_operations__operation_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "operation_id",
            "required": true,
            "schema": {
              "title": "Operation Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Operation",
        "tags": [
          "payments"
        ]
      }
    },
...
    "/api/v1/payment_intents": {
      "post": {
        "operationId": "post_payment_intent_api_v1_payment_intents_post",
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaymentIntentCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Payment Intent",
        "tags": [
          "payments"
        ]
      }
    },
...
    "/api/v1/payment_intents/{payment_intent_id}": {
      "get": {
        "operationId": "get_payment_intent_api_v1_payment_intents__payment_intent_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "payment_intent_id",
            "required": true,
            "schema": {
              "title": "Payment Intent Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Payment Intent",
        "tags": [
          "payments"
        ]
      }
    },
    "/api/v1/payment_intents/{payment_intent_id}/authorize": {
      "post": {
        "operationId": "post_authorize_api_v1_payment_intents__payment_intent_id__authorize_post",
        "parameters": [
          {
            "in": "path",
            "name": "payment_intent_id",
            "required": true,
            "schema": {
              "title": "Payment Intent Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmptyCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Authorize",
        "tags": [
          "payments"
        ]
      }
    },
    "/api/v1/payment_intents/{payment_intent_id}/capture": {
      "post": {
        "operationId": "post_capture_api_v1_payment_intents__payment_intent_id__capture_post",
        "parameters": [
          {
            "in": "path",
            "name": "payment_intent_id",
            "required": true,
            "schema": {
              "title": "Payment Intent Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmptyCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Capture",
        "tags": [
          "payments"
        ]
      }
    },
    "/api/v1/payment_intents/{payment_intent_id}/refunds": {
      "post": {
        "operationId": "post_refund_api_v1_payment_intents__payment_intent_id__refunds_post",
        "parameters": [
          {
            "in": "path",
            "name": "payment_intent_id",
            "required": true,
            "schema": {
              "title": "Payment Intent Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefundCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Refund",
        "tags": [
          "payments"
        ]
      }
    },
...
  }
}